# website/static/openapi.json
{
  "components": {
    "schemas": {
...
      "BulkDnsRecordCreate": {
        "properties": {
          "records": {
            "items": {
              "$ref": "#/components/schemas/DnsRecordCreate"
            },
            "title": "Records",
            "type": "array"
          }
        },
        "required": [
          "records"
        ],
        "title": "BulkDnsRecordCreate",
        "type": "object"
      },
...
      "DnsDiscoveryResponse": {
        "properties": {
          "domain_name": {
            "title": "Domain Name",
            "type": "string"
          },
          "queried_types": {
            "items": {
              "type": "string"
            },
            "title": "Queried Types",
            "type": "array"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/DnsDiscoveryRecord"
            },
            "title": "Records",
            "type": "array"
          },
          "warnings": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "domain_name",
          "records",
          "queried_types"
        ],
        "title": "DnsDiscoveryResponse",
        "type": "object"
      },
      "DnsImportApplyRequest": {
        "properties": {
          "confirm_replace": {
            "default": false,
            "title": "Confirm Replace",
            "type": "boolean"
          },
          "expected_revision": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Expected Revision",
            "type": "string"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/DnsRecordCreate"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Records",
            "type": "array"
          }
        },
        "required": [
          "records",
          "mode",
          "expected_revision"
        ],
        "title": "DnsImportApplyRequest",
        "type": "object"
      },
      "DnsImportPreviewRequest": {
        "properties": {
          "records": {
            "items": {
              "$ref": "#/components/schemas/DnsRecordCreate"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Records",
            "type": "array"
          }
        },
        "required": [
          "records"
        ],
        "title": "DnsImportPreviewRequest",
        "type": "object"
      },
      "DnsImportPreviewResponse": {
        "properties": {
          "additions": {
            "items": {
              "$ref": "#/components/schemas/DnsRecordCreate"
            },
            "title": "Additions",
            "type": "array"
          },
          "conflicts": {
            "items": {
              "$ref": "#/components/schemas/DnsRecordCreate"
            },
            "title": "Conflicts",
            "type": "array"
          },
          "normalized_records": {
            "items": {
              "$ref": "#/components/schemas/DnsRecordCreate"
            },
            "title": "Normalized Records",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          },
          "unchanged": {
            "items": {
              "$ref": "#/components/schemas/DnsRecordCreate"
            },
            "title": "Unchanged",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "additions",
          "unchanged",
          "conflicts",
          "normalized_records",
          "revision"
        ],
        "title": "DnsImportPreviewResponse",
        "type": "object"
      },
      "DnsRecordCreate": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "content": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "name": {
            "maxLength": 253,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "priority": {
            "anyOf": [
              {
                "maximum": 65535.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "proxied": {
            "default": false,
            "title": "Proxied",
            "type": "boolean"
          },
          "record_type": {
            "title": "Record Type",
            "type": "string"
          },
          "ttl": {
            "default": 3600,
            "title": "Ttl",
            "type": "integer"
          }
        },
        "required": [
          "record_type",
          "name",
          "content"
        ],
        "title": "DnsRecordCreate",
        "type": "object"
      },
...
      "DnsTemplateApply": {
        "properties": {
          "params": {
            "additionalProperties": {
              "type": "string"
            },
            "default": {},
            "title": "Params",
            "type": "object"
          },
          "template": {
            "title": "Template",
            "type": "string"
          }
        },
        "required": [
          "template"
        ],
        "title": "DnsTemplateApply",
        "type": "object"
      },
      "DnsZoneExport": {
        "properties": {
          "zone_file": {
            "title": "Zone File",
            "type": "string"
          },
          "zone_name": {
            "title": "Zone Name",
            "type": "string"
          }
        },
        "required": [
          "zone_name",
          "zone_file"
        ],
        "title": "DnsZoneExport",
        "type": "object"
      },
      "DnsZoneResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "default_ttl": {
            "title": "Default Ttl",
            "type": "integer"
          },
          "dnssec_enabled": {
            "title": "Dnssec Enabled",
            "type": "boolean"
          },
          "domain_id": {
            "format": "uuid",
            "title": "Domain Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "primary_ns": {
            "title": "Primary Ns",
            "type": "string"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/DnsRecordResponse"
            },
            "title": "Records",
            "type": "array"
          },
          "serial": {
            "title": "Serial",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "zone_name": {
            "title": "Zone Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "domain_id",
          "zone_name",
          "primary_ns",
          "serial",
          "default_ttl",
          "dnssec_enabled",
          "records",
          "created_at",
          "updated_at"
        ],
        "title": "DnsZoneResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "OpenDomain API. Browser authentication uses email verification, TOTP MFA, and an opaque HttpOnly session cookie. New registrar and payment features remain sandbox-gated until provider approval and production readiness checks pass.",
    "title": "OpenDomain",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/domains/{domain_id}/dns/": {
      "get": {
        "operationId": "get_zone_api_v1_domains__domain_id__dns__get",
        "parameters": [
          {
            "in": "path",
            "name": "domain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Domain Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "opendomain_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opendomain Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DnsZoneResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Zone",
        "tags": [
          "dns"
        ]
      }
    },
    "/api/v1/domains/{domain_id}/dns/discovery": {
      "post": {
        "description": "Discover supported public apex DNS records for an owned domain.",
        "operationId": "discover_dns_api_v1_domains__domain_id__dns_discovery_post",
        "parameters": [
          {
            "in": "path",
            "name": "domain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Domain Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "opendomain_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opendomain Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DnsDiscoveryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Discover Dns",
        "tags": [
          "dns"
        ]
      }
    },
    "/api/v1/domains/{domain_id}/dns/export": {
      "get": {
        "operationId": "export_zone_api_v1_domains__domain_id__dns_export_get",
        "parameters": [
          {
            "in": "path",
            "name": "domain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Domain Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "opendomain_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opendomain Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DnsZoneExport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export Zone",
        "tags": [
          "dns"
        ]
      }
    },
    "/api/v1/domains/{domain_id}/dns/import": {
      "post": {
        "description": "Legacy BIND import. This appends records; use preview/apply APIs for safe imports.",
        "operationId": "import_zone_api_v1_domains__domain_id__dns_import_post",
        "parameters": [
          {
            "in": "path",
            "name": "domain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Domain Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "zone_file",
            "required": true,
            "schema": {
              "title": "Zone File",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "opendomain_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opendomain Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DnsZoneResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Import Zone",
        "tags": [
          "dns"
        ]
      }
    },
    "/api/v1/domains/{domain_id}/dns/import-apply": {
      "post": {
        "description": "Apply a reviewed discovery/import using explicit merge or replacement semantics.",
        "operationId": "apply_dns_import_api_v1_domains__domain_id__dns_import_apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "domain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Domain Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "opendomain_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opendomain Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DnsImportApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DnsZoneResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Apply Dns Import",
        "tags": [
          "dns"
        ]
      }
    },
    "/api/v1/domains/{domain_id}/dns/import-preview": {
      "post": {
        "operationId": "preview_dns_import_api_v1_domains__domain_id__dns_import_preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "domain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Domain Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "opendomain_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opendomain Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DnsImportPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DnsImportPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Preview Dns Import",
        "tags": [
          "dns"
        ]
      }
    },
    "/api/v1/domains/{domain_id}/dns/records": {
      "post": {
        "operationId": "create_record_api_v1_domains__domain_id__dns_records_post",
        "parameters": [
          {
            "in": "path",
            "name": "domain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Domain Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "opendomain_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opendomain Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DnsRecordCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DnsRecordResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Record",
        "tags": [
          "dns"
        ]
      }
    },
    "/api/v1/domains/{domain_id}/dns/records/bulk": {
      "post": {
        "operationId": "create_records_bulk_api_v1_domains__domain_id__dns_records_bulk_post",
        "parameters": [
          {
            "in": "path",
            "name": "domain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Domain Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "opendomain_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opendomain Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDnsRecordCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DnsRecordResponse"
                  },
                  "title": "Response Create Records Bulk Api V1 Domains  Domain Id  Dns Records Bulk Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Records Bulk",
        "tags": [
          "dns"
        ]
      }
    },
    "/api/v1/domains/{domain_id}/dns/records/{record_id}": {
      "delete": {
        "operationId": "delete_record_api_v1_domains__domain_id__dns_records__record_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "domain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Domain Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Record Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "opendomain_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opendomain Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Record",
        "tags": [
          "dns"
        ]
      },
      "patch": {
        "operationId": "update_record_api_v1_domains__domain_id__dns_records__record_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "domain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Domain Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Record Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "opendomain_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opendomain Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DnsRecordUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DnsRecordResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Record",
        "tags": [
          "dns"
        ]
      }
    },
    "/api/v1/domains/{domain_id}/dns/templates": {
      "get": {
        "operationId": "list_templates_api_v1_domains__domain_id__dns_templates_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Templates",
        "tags": [
          "dns"
        ]
      },
      "post": {
        "operationId": "apply_template_api_v1_domains__domain_id__dns_templates_post",
        "parameters": [
          {
            "in": "path",
            "name": "domain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Domain Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "opendomain_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opendomain Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DnsTemplateApply"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DnsZoneResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Apply Template",
        "tags": [
          "dns"
        ]
      }
    },
...
  }
}